requirements.txt
requests
lxml
Pillow
tesserocr
